requirements.txt
flask
python-dotenv
requests
spotipy
key-generator